DataReducer handles firestore access
